# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Event Attendee App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface AdminUser {
  id: number;
  email: string;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  admin: AdminUser;
}

export interface AdminsResponse {
  admins: AdminUser[];
}

export interface CreateAdminRequest {
  email: string;
  /** @minLength 8 */
  password: string;
}

export type DeleteAdminResponseAdmin = {
  id: number;
  email: string;
};

export interface DeleteAdminResponse {
  deleted: boolean;
  admin: DeleteAdminResponseAdmin;
}

export interface RegisterAttendeeRequest {
  /** @minLength 2 */
  fullName: string;
  email: string;
  /** @minLength 7 */
  phoneNumber: string;
  isNewcomer: boolean;
}

export interface Attendee {
  id: number;
  fullName: string;
  email: string;
  phoneNumber: string;
  isNewcomer: boolean;
  createdAt: string;
}

export interface AdminAddAttendeeRequest {
  /** @minLength 2 */
  fullName: string;
  email: string;
  /** @minLength 1 */
  phoneNumber: string;
  isNewcomer: boolean;
  /**
   * Attendance month (1-12). Defaults to current month if omitted.
   * @minimum 1
   * @maximum 12
   */
  month?: number;
  /**
   * Attendance year. Defaults to current year if omitted.
   * @minimum 2000
   * @maximum 2100
   */
  year?: number;
}

export interface AdminAddAttendeeResponse {
  attendee: Attendee;
  /** True if a new attendee row was inserted, false if an existing one was reused. */
  created: boolean;
  /** True if an attendance row was inserted, false if one already existed for (attendee, month, year). */
  attendanceAdded: boolean;
}

export interface ImportAttendeeRow {
  fullName: string;
  email: string;
  phoneNumber: string;
  isNewcomer: boolean;
  /**
   * @minimum 1
   * @maximum 12
   */
  month: number;
  /**
   * @minimum 2000
   * @maximum 2100
   */
  year: number;
}

export interface ImportAttendeesRequest {
  /** @maxItems 5000 */
  rows: ImportAttendeeRow[];
}

export interface ImportError {
  rowNumber: number;
  message: string;
}

export interface ImportAttendeesResponse {
  totalRows: number;
  createdAttendees: number;
  attendancesAdded: number;
  skipped: number;
  errors: ImportError[];
}

export interface DeleteAttendeeResponse {
  deleted: boolean;
  attendee: Attendee;
}

export interface DeleteAttendanceRequest {
  /**
   * @minimum 1
   * @maximum 12
   */
  month: number;
  /**
   * @minimum 2000
   * @maximum 2100
   */
  year: number;
}

export type DeleteAttendanceResponseAttendance = {
  id: number;
  attendeeId: number;
  month: number;
  year: number;
};

export interface DeleteAttendanceResponse {
  deleted: boolean;
  attendance: DeleteAttendanceResponseAttendance;
}

export interface AttendeesResponse {
  attendees: Attendee[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type SmtpSettingsEncryption =
  (typeof SmtpSettingsEncryption)[keyof typeof SmtpSettingsEncryption];

export const SmtpSettingsEncryption = {
  tls: "tls",
  ssl: "ssl",
  none: "none",
} as const;

export interface SmtpSettings {
  host: string;
  port: number;
  username: string;
  encryption: SmtpSettingsEncryption;
  isConfigured: boolean;
}

export type SmtpSettingsRequestEncryption =
  (typeof SmtpSettingsRequestEncryption)[keyof typeof SmtpSettingsRequestEncryption];

export const SmtpSettingsRequestEncryption = {
  tls: "tls",
  ssl: "ssl",
  none: "none",
} as const;

export interface SmtpSettingsRequest {
  host: string;
  port: number;
  username: string;
  password: string;
  encryption: SmtpSettingsRequestEncryption;
}

export type SendEmailRequestTargetGroup =
  (typeof SendEmailRequestTargetGroup)[keyof typeof SendEmailRequestTargetGroup];

export const SendEmailRequestTargetGroup = {
  all: "all",
  newcomers: "newcomers",
  returning: "returning",
} as const;

export interface SendEmailRequest {
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  message: string;
  targetGroup: SendEmailRequestTargetGroup;
  /** Base64-encoded image to embed in the email body */
  imageBase64?: string;
  /** MIME type of the image (e.g. image/png, image/jpeg) */
  imageMimeType?: string;
  /**
   * Only send to attendees who registered in this month (1-12)
   * @minimum 1
   * @maximum 12
   */
  filterMonth?: number;
  /** Only send to attendees who registered in this year */
  filterYear?: number;
}

export interface SmsSettings {
  provider: string;
  senderId: string;
  isConfigured: boolean;
}

export interface SmsSettingsRequest {
  /** @minLength 1 */
  token: string;
  /**
   * @minLength 1
   * @maxLength 11
   */
  senderId: string;
}

export interface SmsTestResponse {
  message: string;
  balance: number;
}

export type SendSmsRequestTargetGroup =
  (typeof SendSmsRequestTargetGroup)[keyof typeof SendSmsRequestTargetGroup];

export const SendSmsRequestTargetGroup = {
  all: "all",
  newcomers: "newcomers",
  returning: "returning",
} as const;

export interface SendSmsRequest {
  /**
   * @minLength 1
   * @maxLength 1600
   */
  message: string;
  targetGroup: SendSmsRequestTargetGroup;
  /**
   * @minimum 1
   * @maximum 12
   */
  filterMonth?: number;
  filterYear?: number;
}

export interface SmsSendError {
  phone: string;
  reason: string;
}

export interface SendSmsResponse {
  successCount: number;
  failedCount: number;
  total: number;
  message: string;
  errors: SmsSendError[];
}

export interface SmsCampaign {
  id: number;
  message: string;
  targetGroup: string;
  filterMonth?: number | null;
  filterYear?: number | null;
  successCount: number;
  failedCount: number;
  total: number;
  sentAt: string;
}

export interface SmsCampaignsResponse {
  campaigns: SmsCampaign[];
}

export interface EmailCampaign {
  id: number;
  subject: string;
  targetGroup: string;
  filterMonth?: number | null;
  filterYear?: number | null;
  successCount: number;
  failedCount: number;
  total: number;
  sentAt: string;
}

export interface EmailCampaignsResponse {
  campaigns: EmailCampaign[];
}

export interface SendEmailResponse {
  successCount: number;
  failedCount: number;
  total: number;
  message: string;
}

export type GetAttendeesParams = {
  /**
   * Search by name or email
   */
  search?: string;
  /**
   * Filter attendees
   */
  filter?: GetAttendeesFilter;
  /**
   * Sort order
   */
  sort?: GetAttendeesSort;
  /**
   * Page number
   * @minimum 1
   */
  page?: number;
  /**
   * Items per page
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * Filter by month (1-12)
   * @minimum 1
   * @maximum 12
   */
  month?: number;
  /**
   * Filter by year (e.g. 2025)
   */
  year?: number;
};

export type GetAttendeesFilter =
  (typeof GetAttendeesFilter)[keyof typeof GetAttendeesFilter];

export const GetAttendeesFilter = {
  all: "all",
  newcomers: "newcomers",
  returning: "returning",
} as const;

export type GetAttendeesSort =
  (typeof GetAttendeesSort)[keyof typeof GetAttendeesSort];

export const GetAttendeesSort = {
  newest: "newest",
  oldest: "oldest",
} as const;

export type ExportAttendeesParams = {
  filter?: ExportAttendeesFilter;
};

export type ExportAttendeesFilter =
  (typeof ExportAttendeesFilter)[keyof typeof ExportAttendeesFilter];

export const ExportAttendeesFilter = {
  all: "all",
  newcomers: "newcomers",
  returning: "returning",
} as const;