# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Event Attendee App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendees:
    post:
      operationId: registerAttendee
      tags: [attendees]
      summary: Register a new attendee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterAttendeeRequest"
      responses:
        "201":
          description: New attendee registered and attendance recorded for the current month
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendee"
        "200":
          description: Returning attendee — attendance recorded for the current month
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendee"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: getAttendees
      tags: [attendees]
      summary: Get attendees with filtering, sorting, pagination
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or email
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, newcomers, returning]
          description: Filter attendees
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest]
          description: Sort order
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Items per page
        - name: month
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 12
          description: Filter by month (1-12)
        - name: year
          in: query
          schema:
            type: integer
          description: Filter by year (e.g. 2025)
      responses:
        "200":
          description: List of attendees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendeesResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendees/admin:
    post:
      operationId: adminAddAttendee
      tags: [attendees]
      summary: Admin add attendee with optional backfilled month/year
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAddAttendeeRequest"
      responses:
        "200":
          description: Attendee added or attendance recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAddAttendeeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendees/import:
    post:
      operationId: importAttendees
      tags: [attendees]
      summary: Bulk import attendees with month/year per row
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportAttendeesRequest"
      responses:
        "200":
          description: Import summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportAttendeesResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendees/{id}:
    delete:
      operationId: deleteAttendee
      tags: [attendees]
      summary: Delete an attendee (cascades to attendances)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attendee deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAttendeeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendees/{id}/attendances:
    delete:
      operationId: deleteAttendance
      tags: [attendees]
      summary: Delete a single attendance for an attendee (specific month/year)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAttendanceRequest"
      responses:
        "200":
          description: Attendance deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAttendanceResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendees/export:
    get:
      operationId: exportAttendees
      tags: [attendees]
      summary: Export attendees as CSV
      security:
        - bearerAuth: []
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, newcomers, returning]
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    RegisterAttendeeRequest:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phoneNumber:
          type: string
          minLength: 7
        isNewcomer:
          type: boolean
      required:
        - fullName
        - email
        - phoneNumber
        - isNewcomer
    Attendee:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        isNewcomer:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - phoneNumber
        - isNewcomer
        - createdAt
    AdminAddAttendeeRequest:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phoneNumber:
          type: string
          minLength: 1
        isNewcomer:
          type: boolean
        month:
          type: integer
          minimum: 1
          maximum: 12
          description: Attendance month (1-12). Defaults to current month if omitted.
        year:
          type: integer
          minimum: 2000
          maximum: 2100
          description: Attendance year. Defaults to current year if omitted.
      required:
        - fullName
        - email
        - phoneNumber
        - isNewcomer
    AdminAddAttendeeResponse:
      type: object
      properties:
        attendee:
          $ref: "#/components/schemas/Attendee"
        created:
          type: boolean
          description: True if a new attendee row was inserted, false if an existing one was reused.
        attendanceAdded:
          type: boolean
          description: True if an attendance row was inserted, false if one already existed for (attendee, month, year).
      required:
        - attendee
        - created
        - attendanceAdded
    ImportAttendeeRow:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        isNewcomer:
          type: boolean
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
          minimum: 2000
          maximum: 2100
      required:
        - fullName
        - email
        - phoneNumber
        - isNewcomer
        - month
        - year
    ImportAttendeesRequest:
      type: object
      properties:
        rows:
          type: array
          maxItems: 5000
          items:
            $ref: "#/components/schemas/ImportAttendeeRow"
      required:
        - rows
    ImportError:
      type: object
      properties:
        rowNumber:
          type: integer
        message:
          type: string
      required:
        - rowNumber
        - message
    ImportAttendeesResponse:
      type: object
      properties:
        totalRows:
          type: integer
        createdAttendees:
          type: integer
        attendancesAdded:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ImportError"
      required:
        - totalRows
        - createdAttendees
        - attendancesAdded
        - skipped
        - errors
    DeleteAttendeeResponse:
      type: object
      properties:
        deleted:
          type: boolean
        attendee:
          $ref: "#/components/schemas/Attendee"
      required:
        - deleted
        - attendee
    DeleteAttendanceRequest:
      type: object
      properties:
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
          minimum: 2000
          maximum: 2100
      required:
        - month
        - year
    DeleteAttendanceResponse:
      type: object
      properties:
        deleted:
          type: boolean
        attendance:
          type: object
          properties:
            id:
              type: integer
            attendeeId:
              type: integer
            month:
              type: integer
            year:
              type: integer
          required:
            - id
            - attendeeId
            - month
            - year
      required:
        - deleted
        - attendance
    AttendeesResponse:
      type: object
      properties:
        attendees:
          type: array
          items:
            $ref: "#/components/schemas/Attendee"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - attendees
        - total
        - page
        - limit
        - totalPages
...