# lib/api-zod/src/generated/types/adminAddAttendeeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Event Attendee App API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminAddAttendeeRequest {
  /** @minLength 2 */
  fullName: string;
  email: string;
  /** @minLength 1 */
  phoneNumber: string;
  isNewcomer: boolean;
  /**
   * Attendance month (1-12). Defaults to current month if omitted.
   * @minimum 1
   * @maximum 12
   */
  month?: number;
  /**
   * Attendance year. Defaults to current year if omitted.
   * @minimum 2000
   * @maximum 2100
   */
  year?: number;
}