# lib/api-zod/src/generated/types/adminAddAttendeeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Event Attendee App API
 * OpenAPI spec version: 0.1.0
 */
import type { Attendee } from "./attendee.js";

export interface AdminAddAttendeeResponse {
  attendee: Attendee;
  /** True if a new attendee row was inserted, false if an existing one was reused. */
  created: boolean;
  /** True if an attendance row was inserted, false if one already existed for (attendee, month, year). */
  attendanceAdded: boolean;
}