# lib/api-zod/src/generated/types/deleteAttendanceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Event Attendee App API
 * OpenAPI spec version: 0.1.0
 */
import type { DeleteAttendanceResponseAttendance } from "./deleteAttendanceResponseAttendance.js";

export interface DeleteAttendanceResponse {
  deleted: boolean;
  attendance: DeleteAttendanceResponseAttendance;
}