# lib/api-zod/src/generated/types/importAttendeesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Event Attendee App API
 * OpenAPI spec version: 0.1.0
 */
import type { ImportError } from "./importError.js";

export interface ImportAttendeesResponse {
  totalRows: number;
  createdAttendees: number;
  attendancesAdded: number;
  skipped: number;
  errors: ImportError[];
}