# lib/api-zod/src/generated/types/importError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Event Attendee App API
 * OpenAPI spec version: 0.1.0
 */

export interface ImportError {
  rowNumber: number;
  message: string;
}